server/requirements.txt
# Core FastAPI dependencies
fastapi==0.115.6
uvicorn[standard]==0.33.0
python-multipart==0.0.20
pydantic==2.10.5
pydantic-settings==2.7.0

# Database (SQLite for development)
sqlalchemy==2.0.36

# Image processing
Pillow==11.1.0
numpy==2.2.1

# Utilities
python-dotenv==1.0.1
aiofiles==24.1.0

# Development dependencies
pytest==8.3.4
pytest-asyncio==0.25.0
